MicroserviceArchitectureProject/author-service/requirements.txt
requests
flask==2.2.2
Werkzeug==2.2.2
mysql-connector-python
python-dotenv